// src/api/generated/models/reviewsControllerGetReviewsByProductV1Params.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * NEST Boilerplate Api Documentation
 * Boilerplate of NEST project Api Documentation to test and review APIs
 * OpenAPI spec version: 1.0
 */
import type { ReviewsControllerGetReviewsByProductV1ReviewSortBy } from "./reviewsControllerGetReviewsByProductV1ReviewSortBy";
import type { ReviewsControllerGetReviewsByProductV1SortOrder } from "./reviewsControllerGetReviewsByProductV1SortOrder";

export type ReviewsControllerGetReviewsByProductV1Params = {
  isPagination?: boolean;
  page?: number;
  limit?: number;
  search?: string;
  sortOrder?: ReviewsControllerGetReviewsByProductV1SortOrder;
  sortBy?: string;
  /**
   * Product ID
   */
  productId?: string;
  /**
   * Sort reviews by
   */
  reviewSortBy?: ReviewsControllerGetReviewsByProductV1ReviewSortBy;
  /**
   * Filter by rating (1-5)
   */
  rating?: number;
};