// src/api/generated/models/reviewsListResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * NEST Boilerplate Api Documentation
 * Boilerplate of NEST project Api Documentation to test and review APIs
 * OpenAPI spec version: 1.0
 */
import type { PaginationMetaResponse } from "./paginationMetaResponse";
import type { ReviewSummary } from "./reviewSummary";
import type { ReviewWithMediaResponse } from "./reviewWithMediaResponse";

export interface ReviewsListResponse {
  data: ReviewWithMediaResponse[];
  summary: ReviewSummary;
  /** @nullable */
  meta: PaginationMetaResponse | null;
}